requirements.txt
docling>=0.5.0
pathlib==1.0.1
pymilvus>=2.5.0
sentence-transformers>=2.2.2
langchain>=0.0.267
regex>=2023.6.3
tqdm>=4.65.0
python-dotenv>=1.0.0
numpy>=1.20.0
torch>=1.10.0
transformers>=4.25.0
scikit-learn>=1.0.0
ijson>=3.2.0
pandas>=1.5.0
pydantic>=1.10.0
fastapi>=0.95.0
uvicorn>=0.20.0
click>=8.1.0
jsonlines>=3.1.0
